User 관련 API 예외처리
- userId가 NaN 일 경우
- 유저 생성 시 동일한 serial_number, phone을 가진 유저가 이미 있을 경우
- 유저 생성 시 필요한 값이 들어오지 않았을 경우
- 유저에게 중복으로 특정 터미널 권한을 부여할 경우

+++ src/api/users/user.controller.js
@@ -1,5 +1,7 @@
 import * as UserService from './user.service.js';
 import { getErrorResponse, getSuccessResponse } from '../../lib/utils.js';
+import { ApiError } from '../../lib/error.js';
+import { apiCode } from '../../lib/api-code.js';
 
 export const getUsers = async (req, res) => {
 	let result;
@@ -21,7 +23,10 @@ export const createUser = async (req, res) => {
 	let { name, phone, serialNumber } = req.body;
 
 	try {
-		let newUser = await UserService.createUser({ name, phone, serialNumber });
+		if (name === undefined || phone === undefined || serialNumber === undefined) {
+			throw new ApiError(apiCode.BAD_REQUEST, 'some value is invalid');
+		}
+		let newUser = await UserService.createUser({ name, phone, serial_number: serialNumber });
 		result = { user: newUser };
 		response = getSuccessResponse(result);
 	} catch (error) {
@@ -33,10 +38,11 @@ export const createUser = async (req, res) => {
 export const modifyUser = async (req, res) => {
 	let result;
 	let response;
-	const userId = req.params.userId;
+	const userId = parseInt(req.params.userId);
 	const { name, phone, serialNumber } = req.body;
 
 	try {
+		if (isNaN(userId)) throw new ApiError(apiCode.BAD_REQUEST, 'userId is NaN');
 		await UserService.modifyUser(userId, { name, phone, serial_number: serialNumber });
 		result = {
 			user: {
@@ -52,9 +58,10 @@ export const modifyUser = async (req, res) => {
 
 export const deleteUser = async (req, res) => {
 	let response;
-	const userId = req.params.userId;
+	const userId = parseInt(req.params.userId);
 
 	try {
+		if (isNaN(userId)) throw new ApiError(apiCode.BAD_REQUEST, 'userId is NaN');
 		await UserService.deleteUser(userId);
 		response = getSuccessResponse({});
 	} catch (error) {
@@ -66,9 +73,10 @@ export const deleteUser = async (req, res) => {
 export const getUserGrants = async (req, res) => {
 	let result;
 	let response;
-	const userId = req.params.userId;
+	const userId = parseInt(req.params.userId);
 
 	try {
+		if (isNaN(userId)) throw new ApiError(apiCode.BAD_REQUEST, 'userId is NaN');
 		const grants = await UserService.getUserGrants(userId);
 		result = {
 			terminals: grants
@@ -83,10 +91,12 @@ export const getUserGrants = async (req, res) => {
 export const addUserGrants = async (req, res) => {
 	let result;
 	let response;
-	const userId = req.params.userId;
-	const { terminalId } = req.body;
+	const userId = parseInt(req.params.userId);
+	let { terminalId } = req.body;
 
 	try {
+		if (isNaN(userId)) throw new ApiError(apiCode.BAD_REQUEST, 'userId is NaN');
+		if (isNaN(parseInt(terminalId))) throw new ApiError(apiCode.BAD_REQUEST, 'terminalId is NaN');
 		const newGrant = await UserService.addUserGrants(userId, terminalId);
 		result = {
 			terminal: newGrant
