+++ src/api/users/user.service.js
@@ -1,6 +1,9 @@
 import { Users } from "../../models/users.model.js";
 import { Terminals } from "../../models/terminals.model.js";
 import { Grants } from "../../models/grants.model.js";
+import { ApiError } from "../../lib/error";
+import { apiCode } from "../../lib/api-code.js";
+import { Op } from "sequelize";
 
 export const getUsers = async () => {
 	const users = await Users.getUsers();
@@ -8,19 +11,35 @@ export const getUsers = async () => {
 }
 
 export const createUser = async (user) => {
+	const user = await Users.findOne({
+		where: {
+			[Op.or]: [{
+				serial_number: user.serial_number
+			}, {
+				phone: user.phone
+			}]
+		}
+	});
+	if (user) throw new ApiError(apiCode.CONFLICT, 'user already exist');
 	const newUser = await Users.createUser(user);
 	return newUser;
 }
 
-export const modifyUser = async (userId, user) => {
-	await Users.updateUser(userId, user)
+export const modifyUser = async (userId, param) => {
+	const user = await Users.getUser(userId);
+	if (!user) throw new ApiError(apiCode.BAD_REQUEST, `The user(${userId}) does not exist`);
+	await Users.updateUser(userId, param)
 }
 
 export const deleteUser = async (userId) => {
+	const user = await Users.getUser(userId);
+	if (!user) throw new ApiError(apiCode.BAD_REQUEST, `The user(${userId}) does not exist`);
 	await Users.deleteUser(userId);
 }
 
 export const getUserGrants = async (userId) => {
+	const user = await Users.getUser(userId);
+	if (!user) throw new ApiError(apiCode.NOT_FOUND, `The user(${userId}) does not exist`)
 	const grants = await Users.getUserGrants(userId);
 	return grants.map((grant) => {
 		delete grant.terminals.grant;
@@ -29,6 +48,13 @@ export const getUserGrants = async (userId) => {
 }
 
 export const addUserGrants = async (userId, terminalId) => {
+	const user = await Users.getUser(userId);
+	if (!user) throw new ApiError(apiCode.NOT_FOUND, `The user(${userId}) does not exist`)
+	const terminal = await Terminals.getTerminal(terminalId);
+	if (!terminal) throw new ApiError(apiCode.NOT_FOUND, `The terminal(${terminalId}) does not exist`)
+	const grant = await Grants.getGrant(user.id, terminal.id);
+	if (grant) throw new ApiError(apiCode.NOT_FOUND, `The user(${userId}) already has grant for terminal(${terminalId})`);
+
 	await Grants.addGrant(userId, terminalId);
-	return Terminals.getTerminal(terminalId);
+	return terminal;
 }
